update server actions and removed side effects
